+++ semana19/projeto/labook/src/business/entities/user.ts
@@ -14,3 +14,4 @@ export type User = {
     email: string,
     password: string
  }
+
