+++ semana19/projeto/labook/src/business/userBusiness.ts
@@ -1,7 +1,7 @@
-import { insertUser } from "../data/userDataBase";
-import { signupInputDTO } from "./entities/user";
+import { insertUser, selectUserByEmail } from "../data/userDataBase";
+import { signupInputDTO, User } from "./entities/user";
 import { generateToken } from "./services/authenticator";
-import { hash } from "./services/hashManager";
+import { compare, hash } from "./services/hashManager";
 import { generateId } from "./services/idGenerator";
 
 export const businessSignup = async (
@@ -25,6 +25,35 @@ export const businessSignup = async (
      })
 
      const token: string = generateToken({ id })
+
+     return token
+
+}
+
+export const businessLogin = async(
+    email: string,
+    password: string
+) => {
+
+    if (!email || !password) { 
+        throw new Error('"email" and "password" must be provided')
+     }
+
+     const user: User = await selectUserByEmail(email) 
+
+      if (!user) { 
+         throw new Error("Invalid credentials")
+      }
+
+      const passwordIsCorrect: boolean = await compare(password, user.password)
+
+      if (!passwordIsCorrect) { 
+         throw new Error("Invalid credentials")
+      }
+
+      const token: string = generateToken({
+        id: user.id
+     })
 
      return token
