+++ semana19/projeto/labook/src/controller/routes/userRouter.ts
@@ -1,6 +1,7 @@
 import express from "express";
-import { signup } from "../userController";
+import { login, signup } from "../userController";
 
 export const userRouter = express.Router();
 
 userRouter.post('/users', signup);
+userRouter.post("/user", login);
