+++ semana19/projeto/labook/src/data/userDataBase.ts
@@ -13,3 +13,25 @@ export const insertUser = async(
        password: user.password
     })
 }
+
+export const selectUserByEmail = async(
+    email: string
+): Promise<User> => {
+    try {
+
+        const queryResult = await connection("labook_users")
+        .select("*")
+        .where({ email })
+
+        return {
+         id: queryResult[0].id,
+         name: queryResult[0].name,
+         email: queryResult[0].email,
+         password: queryResult[0].password
+        }
+
+    } catch (error) {
+        throw new Error(error.slqMessage || error.message)
+    }
+    
+}
