+++ semana19/projeto/labook/src/index.ts
@@ -21,60 +21,7 @@ app.use(cors())
 /**************************** ENDPOINTS ******************************/
 
 app.use('/signup', userRouter)
-
-
-
-
-app.post('/users/login', async (req: Request, res: Response) => {
-   try {
-      let message = "Success!"
-
-      const { email, password } = req.body
-
-      if (!email || !password) {
-         res.statusCode = 406
-         message = '"email" and "password" must be provided'
-         throw new Error(message)
-      }
-
-      const queryResult: any = await connection("labook_users")
-         .select("*")
-         .where({ email })
-
-      if (!queryResult[0]) {
-         res.statusCode = 401
-         message = "Invalid credentials"
-         throw new Error(message)
-      }
-
-      const user: User = {
-         id: queryResult[0].id,
-         name: queryResult[0].name,
-         email: queryResult[0].email,
-         password: queryResult[0].password
-      }
-
-      const passwordIsCorrect: boolean = await compare(password, user.password)
-
-      if (!passwordIsCorrect) {
-         res.statusCode = 401
-         message = "Invalid credentials"
-         throw new Error(message)
-      }
-
-      const token: string = generateToken({
-         id: user.id
-      })
-
-      res.status(200).send({ message, token })
-
-   } catch (error) {
-      let message = error.sqlMessage || error.message
-      res.statusCode = 400
-
-      res.send({ message })
-   }
-})
+app.use('/login', userRouter) 
 
 app.post('/posts/create', async (req: Request, res: Response) => {
    try {
